new code version

+++ src/main/java/io/github/dataforseo/client/model/DataforseoLabsErrorsResultInfo.java
@@ -229,9 +229,9 @@ public class DataforseoLabsErrorsResultInfo  {
 
   public static final String SERIALIZED_NAME_HTTP_TIME = "http_time";
   @SerializedName(SERIALIZED_NAME_HTTP_TIME)
-  private Integer httpTime;
+  private Double httpTime;
 
-  public DataforseoLabsErrorsResultInfo httpTime(Integer httpTime) {
+  public DataforseoLabsErrorsResultInfo httpTime(Double httpTime) {
     this.httpTime = httpTime;
     return this;
   }
@@ -241,11 +241,11 @@ public class DataforseoLabsErrorsResultInfo  {
    * @return httpTime
    */
   @javax.annotation.Nullable
-  public Integer getHttpTime() {
+  public Double getHttpTime() {
     return httpTime;
   }
 
-  public void setHttpTime(Integer httpTime) {
+  public void setHttpTime(Double httpTime) {
     this.httpTime = httpTime;
   }
